// src/types/enums.ts
export enum QuestionType {
  BOOLEAN = "BOOLEAN",
  TEXT = "TEXT",
  DROPDOWN = "DROPDOWN",
  TAGBOX = "TAGBOX",
  RATING = "RATING",
  EMAIL = "EMAIL",
  PHONE = "PHONE",
  DATE = "DATE",
  COMMENT = "COMMENT",
  MATRIX = "MATRIX",
}

export enum ElementTypes {
  text = "text",
  image = "image",
  shape = "shape",
  video = "video",
  videos = "videos",
  button = "button",
  token = "token",
  survey = "survey",
  question = "question",
}